Move Gui classes to Cli namespace
The Gui classes have been relocated to the Cli namespace from the original Gui namespace. This move is reflected in all affected files, updates have been made to the namespace declarations in each of the corresponding classes. The commit also includes the necessary updates to the "using" directives in the affected files to account for this namespace change.