Add console navigation services and pages to CLI
Added a set of services and pages that allow navigation within the console application. Also created interfaces for these components to facilitate dependency injection. In addition, modified the HostService class to correspondingly use these new functionalities. A new way to present service bus structure as a tree was also created.