Restructure GUI framework and update various files
Removed some unused classes and interfaces from the CLI GUI actions and menus. Transferred several classes to a phase-in GUI framework WIP package. This commit also adds additional functionality and updates to many classes, including restructuring the factory to hold a count of queues and topics of the service bus entities. Added "ConsoleWindow" as a newer version of the menu, supporting navigation and a breadcrumb feature. It serves as an entry point for the new GUI framework. The creation of menus is now more explicit and flexible.