Include ServiceBusConfigurationItems.json in project build
The ServiceBusConfigurationItems.json file has been included in the project build process. A new ItemGroup entry with the file reference and CopyToOutputDirectory attribute set to PreserveNewest was added to the csproj file. Additionally, Program.cs was updated to use this config file and an IServiceBusRepository instance was registered in the service collection.